+++ server/service/src/sync/translations/item.rs
@@ -22,7 +22,7 @@ pub struct LegacyItemRow {
     #[serde(deserialize_with = "empty_str_as_option_string")]
     unit_ID: Option<String>,
     type_of: LegacyItemType,
-    default_pack_size: u32,
+    default_pack_size: f64,
 }
 
 fn to_item_type(type_of: LegacyItemType) -> ItemType {
@@ -68,7 +68,7 @@ impl SyncTranslation for ItemTranslation {
             unit_id: data.unit_ID,
             r#type: to_item_type(data.type_of),
             legacy_record: ordered_simple_json(&sync_record.data)?,
-            default_pack_size: data.default_pack_size as i32,
+            default_pack_size: data.default_pack_size,
             is_active: true,
         };
